fixed endpoints and multiple creation of mappers
